+++ src/IECoreRenderMan/Light.cpp
@@ -122,7 +122,7 @@ Light::Light( const ConstGeometryPrototypePtr &geometryPrototype, const Attribut
 
 {
 	m_allAttributes.SetString( Loader::strings().k_grouping_membership, g_defaultLightGroup );
-	m_allAttributes.Update( attributes->instanceAttributes() );
+	m_allAttributes.Update( attributes->lightInstanceAttributes() );
 
 	m_lightShader = acquireLightShader( attributes );
 	if( !m_lightShader || m_lightShader->id() == riley::LightShaderId::InvalidId() )
@@ -218,7 +218,7 @@ bool Light::attributes( const IECoreScenePreview::Renderer::AttributesInterface 
 		return true;
 	}
 
-	m_allAttributes.Update( renderManAttributes->instanceAttributes() );
+	m_allAttributes.Update( renderManAttributes->lightInstanceAttributes() );
 
 	const Material *material = renderManAttributes->lightMaterial();
 	const riley::LightInstanceResult result = m_session->modifyLightInstance(